requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==2.0.12
click==8.1.8
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
line_profiler==4.2.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==0.13.1
py==1.11.0
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2024.2
requests==2.26.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.41.3
toml==0.10.2
tomli==2.2.1
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.34.0
